+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 #### API
 
+- Schedulers warn if no solutions are inserted into archive (#320)
 - Implement `BanditScheduler` (#299)
 - **Backwards-incompatible:** Implement Scalable CMA-ES Optimizers (#274, #288)
 - Make ribs.emitters.opt public (#281)